# pathology-api/openapi.yaml
openapi: 3.0.3
info:
  title: pathology API
  description: Clinical Data pathology API
  version: 0.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5002
    description: Local development server
paths:
  /FHIR/R4/Bundle:
    post:
      summary: Provide a new test result
      description: Provide a new test result to the pathology API
      operationId: postBundle
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              required:
                - resourceType
                - type
                - entry
              properties:
                resourceType:
                  type: string
                  enum:
                    - Bundle
                  description: FHIR resource type
                type:
                  type: string
                  enum:
                    - document
                    - transaction
                  description: The type of the bundle
                  example: document
                entry:
                  type: array
                  description: Entries in the bundle
                  minItems: 1
                  maxItems: 1
                  items:
                    type: object
                    required:
                      - fullUrl
                      - resource
                    properties:
                      fullUrl:
                        type: string
                        description: URI for resource
                        example: "patient"
                      resource:
                        type: object
                        required:
                          - resourceType
                          - identifier
                        properties:
                          resourceType:
                            type: string
                            description: Type of FHIR resource (Always "Patient")
                            enum:
                              - Patient
                            example: Patient
                          identifier:
                            type: object
                            required:
                              - system
                              - value
                            properties:
                              system:
                                type: string
                                enum:
                                  - "https://fhir.nhs.uk/Id/nhs-number"
                                example: "https://fhir.nhs.uk/Id/nhs-number"
                              value:
                                type: string
                                example: "9999999999"

      responses:
        '200':
          description: Successful response
          content:
            application/fhir+json:
              schema:
                schema:
              type: object
              required:
                - resourceType
                - type
              properties:
                resourceType:
                  type: string
                  enum:
                    - Bundle
                  description: FHIR resource type (always "Bundle")
                meta:
                  type: object
                  description: Metadata about the resource
                  nullable: true
                type:
                  type: string
                  enum:
                    - document
                    - transaction
                  description: The type of the bundle
                identifier:
                  type: object
                  nullable: true
                  description: Persistent identifier for the bundle (UUID)
                  properties:
                    system:
                      type: string
                      format: uri
                    value:
                      type: string
                      format: uuid
                entry:
                  type: array
                  nullable: true
                  description: Entries in the bundle
                  items:
                    type: object
                    required:
                      - fullUrl
                      - resource
                    properties:
                      fullUrl:
                        type: string
                        description: URI for resource
                      resource:
                        type: object
                        required:
                          - resourceType
                        description: The Patient a test result is for
                        properties:
                          resourceType:
                            type: string
                            description: Type of FHIR resource (always "Patient")
                            enum:
                              - Patient
                            example: Patient
        '400':
          description: Invalid request
          content:
            text/plain:
              schema:
                type: string